WodingConfig type
