+++ app/server/src/types.ts
@@ -1,0 +1,4 @@
+export interface WodinConfig {
+    port: number,
+    appsPath: string
+}
