--- app/shared/src/config/BasicConfig.ts
@@ -1,3 +1,0 @@
-export interface BasicConfig {
-    basicProp: string
-}
