--- app/shared/src/config/FitConfig.ts
@@ -1,3 +1,0 @@
-export interface FitConfig {
-    fitProp: string
-}
