--- app/shared/src/config/StochasticConfig.ts
@@ -1,3 +1,0 @@
-export interface StochasticConfig {
-    stochasticProp: string
-}
