+++ app/static/src/app/responseTypes.ts
@@ -1,0 +1,11 @@
+export interface BasicConfig {
+    basicProp: string
+}
+
+export interface FitConfig {
+    fitProp: string
+}
+
+export interface StochasticConfig {
+    stochasticProp: string
+}
